refactor(wandb): consolidate import
